fix(rewrite): macro-expanded functions inside impl blocks now instrumented (#643)
Extract impl_prefix_at() into naming.rs as shared helper. Both resolve
(function discovery) and rewriter (guard injection) use the same walk to
find the impl context at a macro call site.

Previously the rewriter called qualified_name_for_fn on standalone
expanded text, producing bare names ("method") that didn't match the
resolve's qualified names ("S::method"). The measured HashMap lookup
failed silently and the function was not instrumented.

Closes #643
